Добавил мои адреса и Макара. Поменял зум у карты и немного изменил структуру проекта
