+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { Layout } from "./Elements/Layout";
-import { TestPage } from "./pages/TestPage";
+import { TeamPage } from "./pages/TeamPage";
 import { useState } from 'react';
 import { ProjectsPage } from "./pages/ProjectsPage";
 
@@ -12,7 +12,7 @@ export function App() {
         <BrowserRouter>
             <Layout theme={theme} setTheme={setTheme}>
                 <Routes>
-                    <Route path="/" element={<TestPage theme={theme} />} />
+                    <Route path="/" element={<TeamPage theme={theme} />} />
                     <Route path="/projects_page" element={<ProjectsPage theme={theme} />} />
                 </Routes>
             </Layout>
