+++ src/components/YandexApi/marksCoords.js
@@ -1,0 +1,50 @@
+import { MAKAR } from '../../static/bio/makarBio';
+import { GRISHA } from '../../static/bio/grisha';
+
+export const COORDS = [
+  {
+    owner: MAKAR,
+    points: [
+      {
+        id: 1000,
+        title: 'Алексеевская улица, 13',
+        coordinates: [56.322934, 44.004343],
+        zoom: 18,
+      },
+      {
+        id: 1002,
+        title: 'улица Стрелка, 21Ж',
+        coordinates: [56.334996, 43.974905],
+        zoom: 18,
+      },
+      {
+        id: 1003,
+        title: 'Заречный бульвар',
+        coordinates: [56.285656, 43.929539],
+      },
+    ],
+  },
+  {
+    owner: GRISHA,
+    points: [
+      {
+        id: 1004,
+        title: 'улица Костякова, 12с9',
+        coordinates: [55.815689, 37.570098],
+        zoom: 18,
+      },
+      {
+        id: 1005,
+        title: 'Тихвинский переулок, 3',
+        coordinates: [55.787609, 37.596044],
+        zoom: 18,
+      },
+      {
+        id: 1005,
+        title: 'парк Дубки',
+        coordinates: [55.82041, 37.566661],
+        zoom: 18,
+      },
+    ],
+  },
+];
