+++ src/pages/TeamPage/index.js
@@ -1,0 +1,1 @@
+export { TeamPage } from "./TeamPage";
