--- src/pages/TestPage/index.js
@@ -1,1 +1,0 @@
-export { TestPage } from "./TestPage";
